requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
supabase>=2.4.0
